% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigexParam.R
\name{addParam,SigexParam,SigexParamTS,GCD-method}
\alias{addParam,SigexParam,SigexParamTS,GCD-method}
\title{addParam}
\usage{
\S4method{addParam}{SigexParam,SigexParamTS,GCD}(object, ts_param, Sigma)
}
\value{
Returns \code{object} invisibly.
}
\description{
Add parameters for a latent component to the structure.
}
\section{Slots}{

\describe{
\item{\code{object}}{A \code{SigexParam} object.}

\item{\code{ts_param}}{A \code{SigexParamTS} object.}

\item{\code{Sigma}}{A \code{GCD} object which represents a \eqn{N \times N}
covariance matrix.}
}}

