% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigexParam.R
\name{asSigexParam,list,list-method}
\alias{asSigexParam,list,list-method}
\title{asSigexParam}
\usage{
\S4method{asSigexParam}{list,list}(paramObject, mdlObject)
}
\value{
A \code{SigexParam}.
}
\description{
Construct a \code{SigexParam} from a \code{sigex} data structure.
}
\details{
Note that \code{mdlObject} is needed to interpret the elements of
\code{paramObject}.
}
\section{Slots}{

\describe{
\item{\code{paramObject}}{A \code{sigex} data structure representing parameters.}

\item{\code{mdlObject}}{A \code{sigex} data structure representing a model.}
}}

