% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{asSigexParam}
\alias{asSigexParam}
\title{asSigexParam}
\usage{
asSigexParam(paramObject, mdlObject)
}
\value{
A \code{SigexParam}.
}
\description{
Construct a \code{SigexParam} from a \code{sigex} data structure.
}
\details{
Note that \code{mdlObject} is needed to interpret the elements of
\code{paramObject}.

AR asks: does this need to be generic?
}
\section{Slots}{

\describe{
\item{\code{paramObject}}{A \code{sigex} data structure representing parameters.}

\item{\code{mdlObject}}{A \code{sigex} data structure representing a model.}
}}

