% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigexParam.R
\name{coef,SigexParam-method}
\alias{coef,SigexParam-method}
\title{coef}
\usage{
\S4method{coef}{SigexParam}(object)
}
\description{
Print elements of the given \code{SigexParam}. (TBD: \code{show} also
does this... maybe this can do something more like other R \code{coef}
methods?)
}
\section{Slots}{

\describe{
\item{\code{object}}{A \code{SigexParam} object.}
}}

