% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigexParam.R
\name{component,SigexParam,numeric-method}
\alias{component,SigexParam,numeric-method}
\title{component}
\usage{
\S4method{component}{SigexParam,numeric}(object, index)
}
\description{
Retrieve a specified component of the \code{SigexParam} and return a new
\code{SigexParam} with only this component.
}
\section{Slots}{

\describe{
\item{\code{object}}{A \code{SigexParam} object.}

\item{\code{index}}{The desired index in \eqn{\{ 1, \ldots, K \}}.}
}}

