% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigexParam.R
\name{setRegParam,SigexParam,numeric-method}
\alias{setRegParam,SigexParam,numeric-method}
\title{setRegParam}
\usage{
\S4method{setRegParam}{SigexParam,numeric}(object, beta)
}
\value{
Returns \code{object} invisibly.
}
\description{
Set the vector of regression coefficients in a \code{SigexParam}.
}
\section{Slots}{

\describe{
\item{\code{object}}{A \code{SigexParam} object.}

\item{\code{beta}}{A vector of regression coefficients.}
}}

