% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigexParam.R
\name{show,SigexParam-method}
\alias{show,SigexParam-method}
\title{show}
\usage{
\S4method{show}{SigexParam}(object)
}
\description{
\code{show} method for \code{SigexParam}.
}
\section{Slots}{

\describe{
\item{\code{object}}{A \code{SigexParam} object.}
}}

