% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SigexFit-class}
\alias{SigexFit-class}
\title{SigexFit}
\description{
Result from Sigex fit via likelihood maximization.
}
\section{Slots}{

\describe{
\item{\code{optimOut}}{Output from the optimizer (\link{optim}).}

\item{\code{param}}{A \code{SigexParam} containing the result.}

\item{\code{data}}{The data matrix}

\item{\code{model}}{The \code{SigexModel} used to define the likelihood.}
}}

