% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SigexModel-class}
\alias{SigexModel-class}
\title{SigexModel}
\description{
Model for an \eqn{N} dimensional series with \eqn{K} latent components.
}
\section{Slots}{

\describe{
\item{\code{N}}{The dimension of the series.}

\item{\code{mdl}}{TBD}

\item{\code{components}}{TBD}
}}

