% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SigexModelComponent-class}
\alias{SigexModelComponent-class}
\title{SigexModelComponent}
\description{
An abstract base class for model components.
}
\section{Slots}{

\describe{
\item{\code{model_class}}{A string that uniquely describes the class of model
component for a subclass of this class.}

\item{\code{delta}}{Vector of coefficients of the differencing polynomial.}

\item{\code{epithet}}{A description for the component}
}}

