% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SigexModelComponentARMA-class}
\alias{SigexModelComponentARMA-class}
\title{SigexModelComponentARMA}
\description{
A subclass of SigexModelComponent that represents an \eqn{\text{ARMA}(p,q)}
series or a vector of independent \eqn{\text{ARMA}(p,q)} series.
}
\section{Slots}{

\describe{
\item{\code{p}}{The order of the AR component. A non-negative integer.}

\item{\code{q}}{The order of the MA component. A non-negative integer.}
}}

