% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SigexModelComponentVARMA-class}
\alias{SigexModelComponentVARMA-class}
\title{SigexModelComponentVARMA}
\description{
A subclass of SigexModelComponent that represents a \eqn{\text{VARMA}(p,q)} process.
}
\section{Slots}{

\describe{
\item{\code{p}}{The order of the AR component. A non-negative integer.}

\item{\code{q}}{The order of the MA component. A non-negative integer.}
}}

