% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SigexParam-class}
\alias{SigexParam-class}
\title{SigexParam}
\description{
Parameters for an \eqn{N} dimensional series with \eqn{K} latent components.
}
\section{Slots}{

\describe{
\item{\code{N}}{The dimension of the series.}

\item{\code{gcds}}{A list of \eqn{K} \code{GCD} objects which represent the
covariance for each component.}

\item{\code{ts_params}}{A list of \eqn{K} time series parameter objects (of type
\code{SigexParamTS}).}

\item{\code{reg_param}}{A vector of regression parameters.}
}}

