% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigexParam.R
\name{SigexParam}
\alias{SigexParam}
\title{SigexParam}
\usage{
SigexParam(N)
}
\description{
Construct a new \code{SigexParam} for an \code{N}-dimensional series, with
no latent components or regression model.
}
\section{Slots}{

\describe{
\item{\code{N}}{A \code{SigexParam} object.}
}}

