% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SigexParamVARMA-class}
\alias{SigexParamVARMA-class}
\title{SigexParamVARMA}
\description{
Parameters structure for an \\eqn{N}-dimensional series with independent
\eqn{\text{VARMA}(p,q)} coordinates.
}
\section{Slots}{

\describe{
\item{\code{ar}}{An \eqn{N \times N \times p} array, where slice \code{[,,l]}
represents the \eqn{l}th AR coefficient matrix.}

\item{\code{ma}}{An \eqn{N \times N \times p} array, where slice \code{[,,l]}
represents the \eqn{l}th MA coefficient matrix.}
}}

