% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigexParam.R
\name{to_sigex,SigexParam-method}
\alias{to_sigex,SigexParam-method}
\title{to_sigex}
\usage{
\S4method{to_sigex}{SigexParam}(object)
}
\value{
A \code{sigex} data structure.
}
\description{
Construct a \code{sigex} data structure for parameters from a
\code{SigexParam}.
}
\section{Slots}{

\describe{
\item{\code{object}}{A \code{SigexParam} object.}
}}

