% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{to_sigex}
\alias{to_sigex}
\title{to_sigex}
\usage{
to_sigex(object)
}
\value{
A \code{sigex} data structure.
}
\description{
Construct a \code{sigex} data structure from the given argument.
}
\section{Slots}{

\describe{
\item{\code{object}}{Argument to convert.}
}}

